Create coffee.csv
The grocery store discontinued the coffee we like so we're using Design of Experiments to figure out which one we should regularly buy next. In addition to varying the coffee type, there are two of us who brew the coffee (NR & MB), we use two different kinds of milk (from the store and from the farm), we vary the number of scoops we put in, and the brew technique (DR vs. default). There is a relationship between "who" and "settings" because MB always uses DR mode, and NR always uses default mode.